op-proposer: Refactor source into a separate package.
