+++ op-e2e/actions/helpers/l2_proposer.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/ethereum-optimism/optimism/op-e2e/config"
+	"github.com/ethereum-optimism/optimism/op-proposer/proposer/source"
 
 	"github.com/ethereum-optimism/optimism/op-service/sources/batching"
 	"github.com/ethereum/go-ethereum"
@@ -109,7 +110,7 @@ func NewL2Proposer(t Testing, log log.Logger, cfg *ProposerCfg, l1 *ethclient.Cl
 		Txmgr:          fakeTxMgr{from: crypto.PubkeyToAddress(cfg.ProposerKey.PublicKey)},
 		L1Client:       l1,
 		Multicaller:    batching.NewMultiCaller(l1.Client(), batching.DefaultBatchSize),
-		ProposalSource: proposer.NewRollupProposalSource(rollupProvider),
+		ProposalSource: source.NewRollupProposalSource(rollupProvider),
 	}
 
 	dr, err := proposer.NewL2OutputSubmitter(driverSetup)
@@ -224,20 +225,20 @@ func toCallArg(msg ethereum.CallMsg) interface{} {
 	return arg
 }
 
-func (p *L2Proposer) fetchNextOutput(t Testing) (proposer.Proposal, bool, error) {
+func (p *L2Proposer) fetchNextOutput(t Testing) (source.Proposal, bool, error) {
 	if p.allocType.UsesProofs() {
 		output, shouldPropose, err := p.driver.FetchDGFOutput(t.Ctx())
 		if err != nil || !shouldPropose {
-			return proposer.Proposal{}, false, err
+			return source.Proposal{}, false, err
 		}
 		encodedBlockNumber := make([]byte, 32)
 		binary.BigEndian.PutUint64(encodedBlockNumber[24:], output.SequenceNum)
 		game, err := p.disputeGameFactory.Games(&bind.CallOpts{}, p.driver.Cfg.DisputeGameType, output.Root, encodedBlockNumber)
 		if err != nil {
-			return proposer.Proposal{}, false, err
+			return source.Proposal{}, false, err
 		}
 		if game.Timestamp != 0 {
-			return proposer.Proposal{}, false, nil
+			return source.Proposal{}, false, nil
 		}
 
 		return output, true, nil
