+++ op-proposer/proposer/driver.go
@@ -12,6 +12,7 @@ import (
 	"github.com/ethereum-optimism/optimism/op-proposer/bindings"
 	"github.com/ethereum-optimism/optimism/op-proposer/contracts"
 	"github.com/ethereum-optimism/optimism/op-proposer/metrics"
+	"github.com/ethereum-optimism/optimism/op-proposer/proposer/source"
 	"github.com/ethereum-optimism/optimism/op-service/dial"
 	"github.com/ethereum-optimism/optimism/op-service/eth"
 	"github.com/ethereum-optimism/optimism/op-service/sources/batching"
@@ -65,7 +66,7 @@ type DriverSetup struct {
 	Multicaller *batching.MultiCaller
 
 	// ProposalSource retrieves the proposal data to submit
-	ProposalSource ProposalSource
+	ProposalSource source.ProposalSource
 }
 
 // L2OutputSubmitter is responsible for proposing outputs
@@ -209,9 +210,9 @@ func (l *L2OutputSubmitter) StopL2OutputSubmitting() error {
 // It returns the output to propose, and whether the proposal should be submitted at all.
 // The passed context is expected to be a lifecycle context. A network timeout
 // context will be derived from it.
-func (l *L2OutputSubmitter) FetchL2OOOutput(ctx context.Context) (Proposal, bool, error) {
+func (l *L2OutputSubmitter) FetchL2OOOutput(ctx context.Context) (source.Proposal, bool, error) {
 	if l.l2ooContract == nil {
-		return Proposal{}, false, fmt.Errorf("L2OutputOracle contract not set, cannot fetch next output info")
+		return source.Proposal{}, false, fmt.Errorf("L2OutputOracle contract not set, cannot fetch next output info")
 	}
 
 	cCtx, cancel := context.WithTimeout(ctx, l.Cfg.NetworkTimeout)
@@ -222,24 +223,24 @@ func (l *L2OutputSubmitter) FetchL2OOOutput(ctx context.Context) (Proposal, bool
 	}
 	nextCheckpointBlockBig, err := l.l2ooContract.NextBlockNumber(callOpts)
 	if err != nil {
-		return Proposal{}, false, fmt.Errorf("querying next block number: %w", err)
+		return source.Proposal{}, false, fmt.Errorf("querying next block number: %w", err)
 	}
 	nextCheckpointBlock := nextCheckpointBlockBig.Uint64()
 	// Fetch the current L2 heads
 	currentBlockNumber, err := l.FetchCurrentBlockNumber(ctx)
 	if err != nil {
-		return Proposal{}, false, err
+		return source.Proposal{}, false, err
 	}
 
 	// Ensure that we do not submit a block in the future
 	if currentBlockNumber < nextCheckpointBlock {
 		l.Log.Debug("Proposer submission interval has not elapsed", "currentBlockNumber", currentBlockNumber, "nextBlockNumber", nextCheckpointBlock)
-		return Proposal{}, false, nil
+		return source.Proposal{}, false, nil
 	}
 
 	output, err := l.FetchOutput(ctx, nextCheckpointBlock)
 	if err != nil {
-		return Proposal{}, false, fmt.Errorf("fetching output: %w", err)
+		return source.Proposal{}, false, fmt.Errorf("fetching output: %w", err)
 	}
 
 	// Always propose if it's part of the Finalized L2 chain. Or if allowed, if it's part of the safe L2 chain.
@@ -259,37 +260,37 @@ func (l *L2OutputSubmitter) FetchL2OOOutput(ctx context.Context) (Proposal, bool
 // a boolean for whether the proposal should be submitted at all.
 // The passed context is expected to be a lifecycle context. A network timeout
 // context will be derived from it.
-func (l *L2OutputSubmitter) FetchDGFOutput(ctx context.Context) (Proposal, bool, error) {
+func (l *L2OutputSubmitter) FetchDGFOutput(ctx context.Context) (source.Proposal, bool, error) {
 	cutoff := time.Now().Add(-l.Cfg.ProposalInterval)
 	proposedRecently, proposalTime, claim, err := l.dgfContract.HasProposedSince(ctx, l.Txmgr.From(), cutoff, l.Cfg.DisputeGameType)
 	if err != nil {
-		return Proposal{}, false, fmt.Errorf("could not check for recent proposal: %w", err)
+		return source.Proposal{}, false, fmt.Errorf("could not check for recent proposal: %w", err)
 	}
 
 	if proposedRecently {
 		l.Log.Debug("Duration since last game not past proposal interval", "duration", time.Since(proposalTime))
-		return Proposal{}, false, nil
+		return source.Proposal{}, false, nil
 	}
 
 	// Fetch the current L2 heads
 	currentBlockNumber, err := l.FetchCurrentBlockNumber(ctx)
 	if err != nil {
-		return Proposal{}, false, fmt.Errorf("could not fetch current block number: %w", err)
+		return source.Proposal{}, false, fmt.Errorf("could not fetch current block number: %w", err)
 	}
 
 	if currentBlockNumber == 0 {
 		l.Log.Info("Skipping proposal for genesis block")
-		return Proposal{}, false, nil
+		return source.Proposal{}, false, nil
 	}
 
 	output, err := l.FetchOutput(ctx, currentBlockNumber)
 	if err != nil {
-		return Proposal{}, false, fmt.Errorf("could not fetch output at current block number %d: %w", currentBlockNumber, err)
+		return source.Proposal{}, false, fmt.Errorf("could not fetch output at current block number %d: %w", currentBlockNumber, err)
 	}
 
 	if claim == output.Root {
 		l.Log.Debug("Skipping proposal: output root unchanged since last proposed game", "last_proposed_root", claim, "output_root", output.Root)
-		return Proposal{}, false, nil
+		return source.Proposal{}, false, nil
 	}
 
 	l.Log.Info("No proposals found for at least proposal interval, submitting proposal now", "proposalInterval", l.Cfg.ProposalInterval)
@@ -312,27 +313,27 @@ func (l *L2OutputSubmitter) FetchCurrentBlockNumber(ctx context.Context) (uint64
 	return status.FinalizedL2, nil
 }
 
-func (l *L2OutputSubmitter) FetchOutput(ctx context.Context, block uint64) (Proposal, error) {
+func (l *L2OutputSubmitter) FetchOutput(ctx context.Context, block uint64) (source.Proposal, error) {
 	output, err := l.ProposalSource.ProposalAtBlock(ctx, block)
 	if err != nil {
-		return Proposal{}, fmt.Errorf("fetching output at block %d: %w", block, err)
+		return source.Proposal{}, fmt.Errorf("fetching output at block %d: %w", block, err)
 	}
 	if output.Version != supportedL2OutputVersion {
-		return Proposal{}, fmt.Errorf("unsupported l2 output version: %v, supported: %v", output.Version, supportedL2OutputVersion)
+		return source.Proposal{}, fmt.Errorf("unsupported l2 output version: %v, supported: %v", output.Version, supportedL2OutputVersion)
 	}
 	if onum := output.SequenceNum; onum != block { // sanity check, e.g. in case of bad RPC caching
-		return Proposal{}, fmt.Errorf("output block number %d mismatches requested %d", output.SequenceNum, block)
+		return source.Proposal{}, fmt.Errorf("output block number %d mismatches requested %d", output.SequenceNum, block)
 	}
 	return output, nil
 }
 
 // ProposeL2OutputTxData creates the transaction data for the ProposeL2Output function
-func (l *L2OutputSubmitter) ProposeL2OutputTxData(output Proposal) ([]byte, error) {
+func (l *L2OutputSubmitter) ProposeL2OutputTxData(output source.Proposal) ([]byte, error) {
 	return proposeL2OutputTxData(l.l2ooABI, output)
 }
 
 // proposeL2OutputTxData creates the transaction data for the ProposeL2Output function
-func proposeL2OutputTxData(abi *abi.ABI, output Proposal) ([]byte, error) {
+func proposeL2OutputTxData(abi *abi.ABI, output source.Proposal) ([]byte, error) {
 	return abi.Pack(
 		"proposeL2Output",
 		output.Root,
@@ -341,7 +342,7 @@ func proposeL2OutputTxData(abi *abi.ABI, output Proposal) ([]byte, error) {
 		new(big.Int).SetUint64(output.CurrentL1.Number))
 }
 
-func (l *L2OutputSubmitter) ProposeL2OutputDGFTxCandidate(ctx context.Context, output Proposal) (txmgr.TxCandidate, error) {
+func (l *L2OutputSubmitter) ProposeL2OutputDGFTxCandidate(ctx context.Context, output source.Proposal) (txmgr.TxCandidate, error) {
 	cCtx, cancel := context.WithTimeout(ctx, l.Cfg.NetworkTimeout)
 	defer cancel()
 	return l.dgfContract.ProposalTx(cCtx, l.Cfg.DisputeGameType, output.Root, output.SequenceNum)
@@ -376,7 +377,7 @@ func (l *L2OutputSubmitter) waitForL1Head(ctx context.Context, blockNum uint64) 
 }
 
 // sendTransaction creates & sends transactions through the underlying transaction manager.
-func (l *L2OutputSubmitter) sendTransaction(ctx context.Context, output Proposal) error {
+func (l *L2OutputSubmitter) sendTransaction(ctx context.Context, output source.Proposal) error {
 	l.Log.Info("Proposing output root", "output", output.Root, "block", output.SequenceNum)
 	var receipt *types.Receipt
 	if l.Cfg.DisputeGameFactoryAddr != nil {
@@ -439,7 +440,7 @@ func (l *L2OutputSubmitter) loop() {
 			// A note on retrying: the outer ticker already runs on a short
 			// poll interval, which has a default value of 6 seconds. So no
 			// retry logic is needed around proposal fetching here.
-			var proposal Proposal
+			var proposal source.Proposal
 			var shouldPropose bool
 			var err error
 			if l.dgfContract == nil {
@@ -481,7 +482,7 @@ func (l *L2OutputSubmitter) waitNodeSync() error {
 	})
 }
 
-func (l *L2OutputSubmitter) proposeOutput(ctx context.Context, output Proposal) {
+func (l *L2OutputSubmitter) proposeOutput(ctx context.Context, output source.Proposal) {
 	cCtx, cancel := context.WithTimeout(ctx, 10*time.Minute)
 	defer cancel()
